Move common test environment setup
Environment that's common for XUnit tests and integration tests has to
be done in `TestMain` function as this is the only way to ensure it's
executed before any `Test*` functions. And since Ginkgo-based
integration tests execute entirely in context of the single `TestAPIs`
function, using its `BeforeSuite` block is too late to affect other
`Test*` functions.

The current setup just ensures all necessary kinds are registered with
the k8s scheme.