Handle cases when OS image is not specified in template
The OS image specification is optional and if not specified provider
controllers will assume sane defaults. It's a problem when upgrading as
when not specified the entire structure is missing from the unstructured
object and we cannot set the required properties.

This populates the missing structure in machine template with empty
maps so that we can later set required properties.